refactor: extract selectDatabase to issue logic
