+++ frontend/src/components/Issue/logic/IssueLogic.ts
@@ -1,5 +1,6 @@
 import { IssueTemplate } from "@/plugins";
 import {
+  Database,
   Issue,
   IssueCreate,
   IssuePatch,
@@ -25,6 +26,7 @@ type IssueLogic = {
   // states
   selectedStage: Ref<Stage | StageCreate>;
   selectedTask: Ref<Task | TaskCreate>;
+  selectedDatabase: Ref<Database | undefined>;
 
   // ui state logic
   isTenantMode: Ref<boolean>;
