+++ frontend/src/components/Issue/logic/base.ts
@@ -21,7 +21,7 @@ import {
   stageSlug,
   taskSlug,
 } from "@/utils";
-import { useIssueStore, useProjectStore } from "@/store";
+import { useDatabaseStore, useIssueStore, useProjectStore } from "@/store";
 import { flattenTaskList, TaskTypeWithStatement } from "./common";
 
 export const useBaseIssueLogic = (params: {
@@ -33,6 +33,7 @@ export const useBaseIssueLogic = (params: {
   const router = useRouter();
   const issueStore = useIssueStore();
   const projectStore = useProjectStore();
+  const databaseStore = useDatabaseStore();
 
   const project = computed((): Project => {
     if (create.value) {
@@ -145,6 +146,17 @@ export const useBaseIssueLogic = (params: {
     return taskList[0];
   });
 
+  const selectedDatabase = computed(() => {
+    if (create.value) {
+      const databaseId = (selectedTask.value as TaskCreate).databaseId;
+      if (databaseId) {
+        return databaseStore.getDatabaseById(databaseId);
+      }
+      return undefined;
+    }
+    return (selectedTask.value as Task).database;
+  });
+
   const isTenantMode = computed((): boolean => {
     if (project.value.tenantMode !== "TENANT") return false;
     const { type } = issue.value;
@@ -203,6 +215,7 @@ export const useBaseIssueLogic = (params: {
     createIssue,
     selectedStage,
     selectedTask,
+    selectedDatabase,
     selectStageOrTask,
     selectTask,
     taskStatusOfStage,
